refactor(frontend): update label dialogs
